backend/pyproject.toml
[project]
name = "saju-master"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "fastapi-pagination>=0.15.12",
    "fastapi[standard]>=0.135.3",
    "openai>=2.30.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
